+++ demo/svelte/src/vite-env.d.ts
@@ -1,2 +1,3 @@
 /// <reference types="svelte" />
 /// <reference types="vite/client" />
+/// <reference types="@spiriit/vite-plugin-svg-spritemap/client" />